chore: move land ownership queries to their own file to organise code more
